requirements.txt
Flask
gunicorn
scikit-learn
numpy
pandas
joblib
openai
requests
sentence-transformers
PyMuPDF
python-docx
python-dotenv
